+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [5.1.1] - 2023-06-29
+
 ### Fixed
 
 - Clone config template repo when `config create -t` is used.
@@ -751,7 +753,8 @@ See the Python Migration Guide for details on how to update your YAML configurat
 - Nearly all command line options removed in favor of YAML equivalents.
 - Completely removed old python project & source code
 
-[Unreleased]: https://github.com/recyclarr/recyclarr/compare/v5.1.0...HEAD
+[Unreleased]: https://github.com/recyclarr/recyclarr/compare/v5.1.1...HEAD
+[5.1.1]: https://github.com/recyclarr/recyclarr/compare/v5.1.0...v5.1.1
 [5.1.0]: https://github.com/recyclarr/recyclarr/compare/v5.0.3...v5.1.0
 [5.0.3]: https://github.com/recyclarr/recyclarr/compare/v5.0.2...v5.0.3
 [5.0.2]: https://github.com/recyclarr/recyclarr/compare/v5.0.1...v5.0.2